Fix 500 error: Add Pillow and typecast numpy types

+++ app.py
@@ -22,8 +22,8 @@ def predict(img):
 
     predictions = model.predict(img_array)
 
-    predicted_class = class_names[np.argmax(predictions[0])]
-    confidence = round(100 * (np.max(predictions[0])), 2)
+    predicted_class = class_names[int(np.argmax(predictions[0]))]
+    confidence = float(round(100 * (np.max(predictions[0])), 2))
     return predicted_class, confidence
 
 # Route to the home page
